+++ src/url_idna.h
@@ -6,8 +6,8 @@
 
 namespace whatwg {
 
-bool IDNToASCII(const char16_t* src, int src_len, simple_buffer<char16_t>& output);
-bool IDNToUnicode(const char* src, int src_len, simple_buffer<char>& output);
+bool IDNToASCII(const char16_t* src, size_t src_len, simple_buffer<char16_t>& output);
+bool IDNToUnicode(const char* src, size_t src_len, simple_buffer<char>& output);
 
 } // namespace whatwg
